+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # hpqtypes-extras-1.16.4.4 (2023-??-??)
 * Switch from `cryptonite` to `crypton`.
+* Make `sqlWhereEqualsAny`, `sqlWhereIn` and `sqlWhereNotIn` prepared-query
+  friendly.
 
 # hpqtypes-extras-1.16.4.3 (2023-06-12)
 * Synchronize timezone of a session with timezone of a database after changing